Fullscreen & Character limit Fixes
Fixes for Raspbian compatability.

"Esc" at fullscreen wasn't registering on Raspbian
